refactor: move Sidenav component to a dedicated sidenav directory
